refactor: 세션 페이지 hooks, view 생성
