+++ frontend/src/components/session/CommonTools.tsx
