+++ frontend/src/components/session/VideoContainer.tsx
@@ -18,6 +18,32 @@ interface VideoContainerProps {
   videoCount: number;
 }
 
+const getVideoLayoutClass = (count: number) => {
+  switch (count) {
+    case 1:
+      return "w-[calc(min(100%,((100vh-140px)*(4/3))))]";
+    case 2:
+      return `w-[calc(min(100%,((100vh-140px)*(2/3))))]
+              sm:w-[calc(min(50%,((100vh-140px)*(4/3))))]
+             `;
+    case 3:
+      return `w-[calc(min(100%,((100vh-140px)*(4/9))))]
+              md:w-[calc(min(50%,((100vh-140px)*(2/3))))]
+              2xl:w-[calc(min(33.3%,((100vh-140px)*(4/3))))] 
+             `;
+    case 4:
+      return `w-[calc(min(100%,((100vh-140px)*(1/3))))]
+              sm:w-[calc(min(100%,((100vh-140px)*(1/3))))]
+              md:w-[calc(min(50%,((100vh-140px)*(2/3))))]
+             `;
+    case 5:
+      return `w-[calc(min(100%,((100vh-140px)*(4/15))))]
+              xs:w-[calc(min(50%,((100vh-140px)*(4/9))))]
+              2xl:w-[calc(min(33.3%,((100vh-140px)*(2/3))))]
+             `;
+  }
+};
+
 const VideoContainer = ({
   nickname,
   isMicOn,
@@ -52,32 +78,6 @@ const VideoContainer = ({
     ) : (
       <BsCameraVideoOff className="text-point-1" />
     );
-  };
-
-  const getVideoLayoutClass = (count: number) => {
-    switch (count) {
-      case 1:
-        return "w-[calc(min(100%,((100vh-140px)*(4/3))))]";
-      case 2:
-        return `w-[calc(min(100%,((100vh-140px)*(2/3))))]
-                sm:w-[calc(min(50%,((100vh-140px)*(4/3))))]
-               `;
-      case 3:
-        return `w-[calc(min(100%,((100vh-140px)*(4/9))))]
-                md:w-[calc(min(50%,((100vh-140px)*(2/3))))]
-                2xl:w-[calc(min(33.3%,((100vh-140px)*(4/3))))] 
-               `;
-      case 4:
-        return `w-[calc(min(100%,((100vh-140px)*(1/3))))]
-                sm:w-[calc(min(100%,((100vh-140px)*(1/3))))]
-                md:w-[calc(min(50%,((100vh-140px)*(2/3))))]
-               `;
-      case 5:
-        return `w-[calc(min(100%,((100vh-140px)*(4/15))))]
-                xs:w-[calc(min(50%,((100vh-140px)*(4/9))))]
-                2xl:w-[calc(min(33.3%,((100vh-140px)*(2/3))))]
-               `;
-    }
   };
 
   return (
