+++ frontend/src/pages/SessionPage/hooks/useMediaDevices.ts
