+++ frontend/src/pages/SessionPage/hooks/useMediaStreamCleanup.ts
