+++ frontend/src/pages/SessionPage/view/SessionHeader.tsx
