+++ frontend/src/pages/SessionPage/view/SidebarContainer.tsx
