+++ frontend/src/pages/SessionPage/view/VideoLayout.tsx
@@ -24,8 +24,8 @@ const VideoLayout = ({
   const videoCount = 1 + peers.length;
 
   return (
-    <div className="w-full flex-1 min-h-0 overflow-hidden">
-      <div className="flex flex-wrap w-full h-full items-center justify-center">
+    <div className="relative w-full flex-1 min-h-0 overflow-hidden">
+      <div className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 flex flex-wrap w-full justify-center">
         <VideoContainer
           nickname={nickname}
           isMicOn={isMicOn}
